Created messaging between microservices about purchases

+++ RecommendationNetwork/Repositories/PurchaseDetailRepository.cs
@@ -29,8 +29,8 @@ namespace RecommendationNetwork.Repositories
                 Id = propertiesConsistOf["Id"].As<int>(),
                 ProductId = propertiesProduct["Id"].As<int>(),
                 PurchaseId = propertiesPurchase["Id"].As<int>(),
-                Quantity = propertiesConsistOf["Quantity"].As<int>(),
-                PriceForOnePiece = propertiesConsistOf["PriceForOnePiece"].As<int>()
+                Quantity = propertiesConsistOf["Quantity"].As<float>(),
+                PriceForOnePiece = propertiesConsistOf["PriceForOnePiece"].As<float>()
             };
 
             return orderResponse;
